requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# server.py: 1
Flask == 1.0.2

# daemon.py: 20
# server.py: 2
cassandra_driver == 3.17.0

# writers/json_writer.py: 1
jsonpickle == 1.0

# csvpo.py: 7,8
# writers/csv_writer.py: 7,8
# writers/json_writer.py: 3,4
jupyter-parser == 0.0a2

# features/psm_features.py: 2,5
numba == 0.42.0

# features/psm_features.py: 1
numpy == 1.16.2

# features/psm_features.py: 6
pyteomics == 4.0.1

# accessors/pride_data.py: 9
requests == 2.21.0